+++ src/test/java/com/vitorbetmann/cleanresmapi/usecases/usertype/CreateUserTypeTest.java
@@ -23,7 +23,7 @@ public class CreateUserTypeTest {
     @InjectMocks
     CreateUserType createUserType;
 
-    Integer mockId = 1;
+    Long mockId = 1L;
     String mockName = "Owner";
 
     @Test
